buf/fixed: make collections inner structs reusable
Move the Inner types of FixedBufRegistry and FixedBufPool
into a plumbing module and expose them within buf::fixed,
to allow code reuse with the upcoming thread-safe collections.
