+++ src/buf/fixed/registry.rs
@@ -1,16 +1,11 @@
-use super::handle::CheckedOutBuf;
-use super::{FixedBuf, FixedBuffers};
+use super::plumbing;
+use super::FixedBuf;
 
 use crate::runtime::driver::WeakHandle;
 use crate::runtime::CONTEXT;
-use libc::{iovec, UIO_MAXIOV};
 use std::cell::RefCell;
-use std::cmp;
 use std::io;
-use std::mem;
-use std::ptr;
 use std::rc::Rc;
-use std::slice;
 
 /// An indexed collection of I/O buffers pre-registered with the kernel.
 ///
@@ -36,7 +31,7 @@ use std::slice;
 /// [`Runtime`]: crate::Runtime
 #[derive(Clone)]
 pub struct FixedBufRegistry {
-    inner: Rc<RefCell<Inner>>,
+    inner: Rc<RefCell<plumbing::Registry>>,
     driver: WeakHandle,
 }
 
@@ -105,7 +100,7 @@ impl FixedBufRegistry {
     /// ```
     pub fn new(bufs: impl IntoIterator<Item = Vec<u8>>) -> Self {
         FixedBufRegistry {
-            inner: Rc::new(RefCell::new(Inner::new(bufs.into_iter()))),
+            inner: Rc::new(RefCell::new(plumbing::Registry::new(bufs.into_iter()))),
             driver: CONTEXT.with(|x| {
                 x.handle()
                     .as_ref()
@@ -175,130 +170,8 @@ impl FixedBufRegistry {
         inner.check_out(index).map(|data| {
             let registry = Rc::clone(&self.inner);
             // Safety: the validity of buffer data is ensured by
-            // Inner::check_out
+            // plumbing::Registry::check_out
             unsafe { FixedBuf::new(registry, data) }
         })
-    }
-}
-
-// Internal state shared by FixedBufRegistry and FixedBuf handles.
-struct Inner {
-    // Pointer to an allocated array of iovec records referencing
-    // the allocated buffers. The number of initialized records is the
-    // same as the length of the states array.
-    raw_bufs: ptr::NonNull<iovec>,
-    // State information on the buffers. Indices in this array correspond to
-    // the indices in the array at raw_bufs.
-    states: Vec<BufState>,
-    // Original capacity of raw_bufs as a Vec.
-    orig_cap: usize,
-}
-
-// State information of a buffer in the registry,
-enum BufState {
-    // The buffer is not in use.
-    // The field records the length of the initialized part.
-    Free { init_len: usize },
-    // The buffer is checked out.
-    // Its data are logically owned by the FixedBuf handle,
-    // which also keeps track of the length of the initialized part.
-    CheckedOut,
-}
-
-impl Inner {
-    fn new(bufs: impl Iterator<Item = Vec<u8>>) -> Self {
-        let bufs = bufs.take(cmp::min(UIO_MAXIOV as usize, u16::MAX as usize));
-        let (size_hint, _) = bufs.size_hint();
-        let mut iovecs = Vec::with_capacity(size_hint);
-        let mut states = Vec::with_capacity(size_hint);
-        for mut buf in bufs {
-            iovecs.push(iovec {
-                iov_base: buf.as_mut_ptr() as *mut _,
-                iov_len: buf.capacity(),
-            });
-            states.push(BufState::Free {
-                init_len: buf.len(),
-            });
-            mem::forget(buf);
-        }
-        debug_assert_eq!(iovecs.len(), states.len());
-
-        // Safety: Vec::as_mut_ptr never returns null
-        let raw_bufs = unsafe { ptr::NonNull::new_unchecked(iovecs.as_mut_ptr()) };
-        let orig_cap = iovecs.capacity();
-        mem::forget(iovecs);
-        Inner {
-            raw_bufs,
-            states,
-            orig_cap,
-        }
-    }
-
-    // If the indexed buffer is free, changes its state to checked out
-    // and returns its data.
-    // If the buffer is already checked out, returns None.
-    fn check_out(&mut self, index: usize) -> Option<CheckedOutBuf> {
-        let state = self.states.get_mut(index)?;
-        let BufState::Free { init_len } = *state else {
-            return None
-        };
-
-        *state = BufState::CheckedOut;
-
-        // Safety: the allocated array under the pointer is valid
-        // for the lifetime of self, the index is inside the array
-        // as checked by Vec::get_mut above, called on the array of
-        // states that has the same length.
-        let iovec = unsafe { self.raw_bufs.as_ptr().add(index).read() };
-        debug_assert!(index <= u16::MAX as usize);
-        Some(CheckedOutBuf {
-            iovec,
-            init_len,
-            index: index as u16,
-        })
-    }
-
-    fn check_in_internal(&mut self, index: u16, init_len: usize) {
-        let state = self
-            .states
-            .get_mut(index as usize)
-            .expect("invalid buffer index");
-        debug_assert!(
-            matches!(state, BufState::CheckedOut),
-            "the buffer must be checked out"
-        );
-        *state = BufState::Free { init_len };
-    }
-}
-
-impl FixedBuffers for Inner {
-    fn iovecs(&self) -> &[iovec] {
-        // Safety: the raw_bufs pointer is valid for the lifetime of self,
-        // the length of the states array is also the length of buffers array
-        // by construction.
-        unsafe { slice::from_raw_parts(self.raw_bufs.as_ptr(), self.states.len()) }
-    }
-
-    unsafe fn check_in(&mut self, index: u16, init_len: usize) {
-        self.check_in_internal(index, init_len)
-    }
-}
-
-impl Drop for Inner {
-    fn drop(&mut self) {
-        let iovecs = unsafe {
-            Vec::from_raw_parts(self.raw_bufs.as_ptr(), self.states.len(), self.orig_cap)
-        };
-        for (i, iovec) in iovecs.iter().enumerate() {
-            match self.states[i] {
-                BufState::Free { init_len } => {
-                    let ptr = iovec.iov_base as *mut u8;
-                    let cap = iovec.iov_len;
-                    let v = unsafe { Vec::from_raw_parts(ptr, init_len, cap) };
-                    mem::drop(v);
-                }
-                BufState::CheckedOut => unreachable!("all buffers must be checked in"),
-            }
-        }
     }
 }
